requirements.txt
pandas
scikit-learn
joblib
PyYAML
pytest
jupyter
matplotlib
